refactoring(integrations): removed inline call of functions, created separate components instead
